requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.1.1
defusedxml==0.7.0rc1
dj-database-url==0.5.0
Django==3.0.7
django-countries==6.1.3
django-crispy-forms==1.9.0
django-js-asset==1.2.2
django-mptt==0.11.0
idna==2.10
Jinja2==2.10.1
jsmin==2.2.2
MarkupSafe==1.1.1
oauthlib==3.1.0
Pillow==7.1.0
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2018.5
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
urllib3==1.25.10
